+++ walkers/src/tiles.rs
@@ -57,6 +57,7 @@ pub trait TilesManager {
     fn at(&mut self, tile_id: TileId) -> Option<Texture>;
     fn attribution(&self) -> Attribution;
     fn tile_size(&self) -> u32;
+    fn available_zoom(&self) -> Vec<u8>;
 }
 
 /// Downloads and keeps cache of the tiles. It must persist between frames.
@@ -140,6 +141,10 @@ impl TilesManager for Tiles {
 
     fn tile_size(&self) -> u32 {
         self.tile_size
+    }
+
+    fn available_zoom(&self) -> Vec<u8> {
+        return Vec::from_iter(0..=19);
     }
 }